Test build step in CI
Recently #5 caused breakage of the automated deployment because of
a syntax typo (fixed in #23). It broke the building of the PDF, which
is not otherwise tested and slow for developers to build it locally
constantly.

CI now has a build job that will build the deployment docker image
just as it does in the real deployment step. Instead of forcing it
to always build "main", allow a different branch to be tested with
`--build-arg GIT_BRANCH=...`.