define "translation freeze" as the basis for the revocation timeline
Version 5 of the policy originally used the term "release/localization
cycle".  @eloquence noted[1] that that this term does not precisely
specify the end of this period: is it when translations for the new
version are merged? or when the new version is actually released?  In
practice, the "translation freeze" coincides with the end of the QA
cycle, so here we define and use the former term explicitly.

[1]: https://github.com/freedomofpress/securedrop-dev-docs/pull/35#discussion_r1137979175