Overhaul the Translations page
This commit overhauls the translations page, in service of the
new continuous localization work that is putting the translation
process for SecureDrop Server in line with the existing
Client workflow.

This commit removes the background sections that are covered
elsewhere in the documentation, and groups the remaining
Weblate-specific items together into a more cohesive section.

It also adds a digram detailing the new continuous localization
workflow (with changes to come), and gives an overview of
responsibilities based on said workflow.