Update SDW Debian package release instructions
This contains an update for two major changes:

== Switch to monorepo ==

All components are now built together via the client repo, so all
changes happen there first and the builder repository is just needed for
localwheels.

We also do a changelog bump in between the last RC and the actual prod
package, so there is no expectation and workarounds to keep the checksum
the same across the two.

== Building in containers ==

Per <https://github.com/freedomofpress/securedrop-engineering/pull/20>,
as long as the package is reproducible, we can build it inside a
container instead of a dispVM. The instructions still require a fresh
clone, but the build script will transparently take care of the
containerization part. An explicit step is added for having another
maintainer check the package is actually reproducible.

== Misc. changes ==

I removed language related to timelines given that we don't actually
want releases to take several weeks, nor have they recently.

Fixes #134.