Document Python and Node.js toolchain usage
Turn the Rust toolchain documentation into a more generic "Language
toolchains" page that also covers Python and Node.js.

For Python discuss that we use the OS-provided versions and update on
that cycle.

For Node.js, we're primarily using it in the context of Electron, so
we'll use the version it provides. Also we're using pnpm.

Fixes #245.